Align census CIDOC exports with CRM domain and range
